Refector db access for comments to a seperate class
