+++ src/main/java/de/samply/bbmri/negotiator/control/researcher/ResearcherQueriesBean.java
@@ -39,6 +39,7 @@ import javax.faces.bean.ManagedBean;
 import javax.faces.bean.ManagedProperty;
 import javax.faces.bean.ViewScoped;
 
+import eu.bbmri.eric.csit.service.negotiator.database.DbUtilComment;
 import eu.bbmri.eric.csit.service.negotiator.database.DbUtilRequest;
 import org.jooq.Record;
 import org.jooq.Result;
@@ -145,7 +146,7 @@ public class ResearcherQueriesBean implements Serializable {
         selectedQuery = query;
 
         try(Config config = ConfigFactory.get()) {
-            comments = DbUtil.getComments(config, selectedQuery.getId(), userBean.getUserId());
+            comments = DbUtilComment.getComments(config, selectedQuery.getId(), userBean.getUserId());
         } catch (SQLException e) {
             e.printStackTrace();
         }
